Move Sequelize configuration files into a subpackage of infra/database folder
Mongoose is another kind of ORM than Sequelize. So we must reorganize `database` folder.
