+++ lib/infrastructure/database/sequelize/models/User.js
